requirements.txt
# --- Airflow Providers ---
apache-airflow-providers-postgres==5.5.0
apache-airflow-providers-google==10.7.0
apache-airflow-providers-apache-beam==5.3.0

# --- PostgreSQL Drivers & ORM ---
psycopg2-binary==2.9.9
SQLAlchemy==1.4.46

# --- Google Cloud SDKs ---
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
google-cloud-core==2.3.3
google-cloud-pubsub==2.15.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0

# --- Data & Time Handling ---
pandas==2.0.3
pytz==2023.3
pyarrow==11.0.0

# --- Utilities ---
requests==2.32.3
python-dotenv==1.0.0
faker==18.9.0

# --- Streaming (Dataflow / Apache Beam) ---
apache-beam[gcp]==2.53.0

# --- DBT Core ---
dbt-core==1.6.6
dbt-postgres==1.6.6
dbt-bigquery==1.6.6

# --- Sender ---
discord_webhook==1.4.1

# --- web scraping ---
selenium==4.11.0
beautifulsoup4==4.13.5

# --- pdf extract ---
sumy==0.11.0
PyMuPDF==1.24.9
nltk==3.9.1
